Extract shared statement-finding logic into parserUtils
Move _findStatementStartLine, _findMatchingOpenBrace, and
_isPrecededByArrow from both C# and JS/TS parsers into a shared
parserUtils.ts module. Both parsers now import from parserUtils
instead of duplicating the identical implementations.

Addresses review feedback from JamesNK about code duplication.
